address comment, remove some redundant tests
